pyproject.toml
[project]
name = "ai-eng-challenge"
version = "0.1.0"
description = "Greeter agent for DEUS Bank"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0",
    "langchain-core>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[tool.uv]
